requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
langchain==0.0.339
langchain_community==0.0.13
google-generativeai==0.3.1
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
pillow==10.1.0
sentence-transformers==2.2.2
requests==2.31.0
email-validator==2.0.0
transformers==4.37.2
